+++ cognite/client/data_classes/data_modeling/data_models.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import json
 from operator import attrgetter
 from typing import TYPE_CHECKING, Any, Generic, Literal, TypeVar, Union
 
@@ -77,14 +76,13 @@ class DataModelApply(DataModelCore):
         if "type" in view_data:
             return ViewId.load(view_data)
         else:
-            return ViewApply.load(view_data)
+            return ViewApply._load(view_data)
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> Self:
-        data = json.loads(resource) if isinstance(resource, str) else resource
-        if "views" in data:
-            data["views"] = [cls._load_view(v) for v in data["views"]] or None
-        return super().load(data)
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Self:
+        if "views" in resource:
+            resource["views"] = [cls._load_view(v) for v in resource["views"]] or None
+        return super()._load(resource)
 
     def dump(self, camel_case: bool = False) -> dict[str, Any]:
         output = super().dump(camel_case)
@@ -138,15 +136,14 @@ class DataModel(DataModelCore, Generic[T_View]):
         if "type" in view_data:
             return ViewId.load(view_data)
         else:
-            return View.load(view_data)
+            return View._load(view_data)
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> Self:
-        data = json.loads(resource) if isinstance(resource, str) else resource
-        if "views" in data:
-            data["views"] = [cls._load_view(v) for v in data["views"]] or None
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> Self:
+        if "views" in resource:
+            resource["views"] = [cls._load_view(v) for v in resource["views"]] or None
 
-        return super().load(data)
+        return super()._load(resource)
 
     def dump(self, camel_case: bool = False) -> dict[str, Any]:
         output = super().dump(camel_case)