+++ cognite/client/data_classes/data_modeling/instances.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import json
 import threading
 from abc import abstractmethod
 from collections import defaultdict
@@ -160,15 +159,14 @@ class InstanceApply(InstanceCore):
         return output
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> Self:
-        data = resource if isinstance(resource, dict) else json.loads(resource)
-        data = convert_all_keys_to_snake_case(data)
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Self:
+        resource = convert_all_keys_to_snake_case(resource)
         if cls is not InstanceApply:
             # NodeApply and EdgeApply does not support instance type
-            data.pop("instance_type", None)
-        instance = cls(**convert_all_keys_to_snake_case(data))
-        if "sources" in data:
-            instance.sources = [NodeOrEdgeData.load(source) for source in data["sources"]]
+            resource.pop("instance_type", None)
+        instance = cls(**convert_all_keys_to_snake_case(resource))
+        if "sources" in resource:
+            instance.sources = [NodeOrEdgeData.load(source) for source in resource["sources"]]
         return instance
 
 
@@ -286,11 +284,10 @@ class Instance(InstanceCore):
         self.properties: Properties = properties or Properties({})
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> Self:
-        data = json.loads(resource) if isinstance(resource, str) else resource
-        if "properties" in data:
-            data["properties"] = Properties.load(data["properties"])
-        return super().load(data)
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> Self:
+        if "properties" in resource:
+            resource["properties"] = Properties.load(resource["properties"])
+        return super()._load(resource)
 
     def dump(self, camel_case: bool = False) -> dict[str, Any]:
         dumped = super().dump(camel_case)
@@ -349,23 +346,21 @@ class InstanceAggregationResult(DataModelingResource):
         self.group = group
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> Self:
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Self:
         """
         Loads an instance from a json string or dictionary.
 
         Args:
-            resource (dict | str): No description.
+            resource (dict): No description.
             cognite_client (CogniteClient | None): No description.
 
         Returns:
             Self: An instance.
 
         """
-        data = json.loads(resource) if isinstance(resource, str) else resource
-
         return cls(
-            aggregates=[AggregatedNumberedValue.load(agg) for agg in data["aggregates"]],
-            group=cast(Dict[str, Union[str, int, float, bool]], data.get("group")),
+            aggregates=[AggregatedNumberedValue.load(agg) for agg in resource["aggregates"]],
+            group=cast(Dict[str, Union[str, int, float, bool]], resource.get("group")),
         )
 
     def dump(self, camel_case: bool = False) -> dict[str, Any]:
@@ -421,10 +416,9 @@ class NodeApply(InstanceApply):
         return output
 
     @classmethod
-    def load(cls, data: dict | str) -> NodeApply:
-        data = json.loads(data) if isinstance(data, str) else data
-        instance = super().load(data)
-        instance.type = DirectRelationReference.load(data["type"]) if "type" in data else None
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> NodeApply:
+        instance = super()._load(resource)
+        instance.type = DirectRelationReference.load(resource["type"]) if "type" in resource else None
         return instance
 
     def as_id(self) -> NodeId:
@@ -497,17 +491,16 @@ class Node(Instance):
         return output
 
     @classmethod
-    def load(cls, data: dict | str) -> Node:
-        data = json.loads(data) if isinstance(data, str) else data
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Node:
         return Node(
-            space=data["space"],
-            external_id=data["externalId"],
-            version=data["version"],
-            last_updated_time=data["lastUpdatedTime"],
-            created_time=data["createdTime"],
-            deleted_time=data.get("deletedTime"),
-            properties=Properties.load(data["properties"]) if "properties" in data else None,
-            type=DirectRelationReference.load(data["type"]) if "type" in data else None,
+            space=resource["space"],
+            external_id=resource["externalId"],
+            version=resource["version"],
+            last_updated_time=resource["lastUpdatedTime"],
+            created_time=resource["createdTime"],
+            deleted_time=resource.get("deletedTime"),
+            properties=Properties.load(resource["properties"]) if "properties" in resource else None,
+            type=DirectRelationReference.load(resource["type"]) if "type" in resource else None,
         )
 
 
@@ -594,13 +587,12 @@ class EdgeApply(InstanceApply):
         return output
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> Self:
-        data = json.loads(resource) if isinstance(resource, str) else resource
-        instance = super().load(data)
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Self:
+        instance = super()._load(resource)
 
-        instance.type = DirectRelationReference.load(data["type"])
-        instance.start_node = DirectRelationReference.load(data["startNode"])
-        instance.end_node = DirectRelationReference.load(data["endNode"])
+        instance.type = DirectRelationReference.load(resource["type"])
+        instance.start_node = DirectRelationReference.load(resource["startNode"])
+        instance.end_node = DirectRelationReference.load(resource["endNode"])
         return instance
 
 
@@ -681,13 +673,12 @@ class Edge(Instance):
         return output
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> Self:
-        data = json.loads(resource) if isinstance(resource, str) else resource
-        instance = super().load(data)
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> Self:
+        instance = super()._load(resource)
 
-        instance.type = DirectRelationReference.load(data["type"])
-        instance.start_node = DirectRelationReference.load(data["startNode"])
-        instance.end_node = DirectRelationReference.load(data["endNode"])
+        instance.type = DirectRelationReference.load(resource["type"])
+        instance.start_node = DirectRelationReference.load(resource["startNode"])
+        instance.end_node = DirectRelationReference.load(resource["endNode"])
         return instance