+++ cognite/client/data_classes/extractionpipelines.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import json
 from typing import TYPE_CHECKING, Any, Sequence, cast
 
 from cognite.client.data_classes._base import (
@@ -47,8 +46,7 @@ class ExtractionPipelineContact(dict):
         return convert_all_keys_to_camel_case(self) if camel_case else dict(self)
 
     @classmethod
-    def load(cls, data: str | dict) -> ExtractionPipelineContact:
-        data = json.loads(data) if isinstance(data, str) else data
+    def load(cls, data: dict[str, Any]) -> ExtractionPipelineContact:
         return cls(
             name=data["name"], email=data["email"], role=data["role"], send_notification=data["sendNotification"]
         )
@@ -122,8 +120,8 @@ class ExtractionPipeline(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> ExtractionPipeline:
-        instance = super().load(resource, cognite_client)
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> ExtractionPipeline:
+        instance = super()._load(resource, cognite_client)
         if instance.contacts:
             instance.contacts = [ExtractionPipelineContact.load(contact) for contact in instance.contacts]
         return instance
@@ -261,8 +259,8 @@ class ExtractionPipelineRun(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> ExtractionPipelineRun:
-        obj = super().load(resource, cognite_client)
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> ExtractionPipelineRun:
+        obj = super()._load(resource, cognite_client)
         # Note: The API ONLY returns IDs, but if they chose to change this, we're ready:
         if isinstance(resource, dict):
             obj.extpipe_external_id = resource.get("externalId")