+++ cognite/client/data_classes/geospatial.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import dataclasses
-import json
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any, List, cast
 
@@ -39,9 +38,7 @@ class FeatureType(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(cls, resource: str | dict[str, Any], cognite_client: CogniteClient | None = None) -> FeatureType:
-        if isinstance(resource, str):
-            return cls.load(json.loads(resource), cognite_client=cognite_client)
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> FeatureType:
         instance = cls(cognite_client=cognite_client)
         for key, value in resource.items():
             snake_case_key = to_snake_case(key)
@@ -92,9 +89,7 @@ class Feature(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(cls, resource: str | dict[str, Any], cognite_client: CogniteClient | None = None) -> Feature:
-        if isinstance(resource, str):
-            return cls.load(json.loads(resource), cognite_client=cognite_client)
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> Feature:
         instance = cls(cognite_client=cognite_client)
         for key, value in resource.items():
             # Keep properties defined in Feature Type as is
@@ -275,9 +270,7 @@ class FeatureAggregate(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(cls, resource: str | dict[str, Any], cognite_client: CogniteClient | None = None) -> FeatureAggregate:
-        if isinstance(resource, str):
-            return cls.load(json.loads(resource), cognite_client=cognite_client)
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> FeatureAggregate:
         instance = cls(cognite_client=cognite_client)
         for key, value in resource.items():
             snake_case_key = to_snake_case(key)
@@ -305,11 +298,7 @@ class CoordinateReferenceSystem(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(
-        cls, resource: str | dict[str, Any], cognite_client: CogniteClient | None = None
-    ) -> CoordinateReferenceSystem:
-        if isinstance(resource, str):
-            return cls.load(json.loads(resource), cognite_client=cognite_client)
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> CoordinateReferenceSystem:
         instance = cls(cognite_client=cognite_client)
         for key, value in resource.items():
             snake_case_key = to_snake_case(key)
@@ -394,11 +383,7 @@ class GeospatialComputedItem(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(
-        cls, resource: str | dict[str, Any], cognite_client: CogniteClient | None = None
-    ) -> GeospatialComputedItem:
-        if isinstance(resource, str):
-            return cls.load(json.loads(resource), cognite_client=cognite_client)
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> GeospatialComputedItem:
         instance = cls(resource=resource, cognite_client=cognite_client)
         for key, value in resource.items():
             snake_case_key = to_snake_case(key)
@@ -422,12 +407,9 @@ class GeospatialComputedResponse(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(
-        cls, resource: str | dict[str, Any], cognite_client: CogniteClient | None = None
-    ) -> GeospatialComputedResponse:
-        resource = json.loads(resource) if isinstance(resource, str) else resource
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> GeospatialComputedResponse:
         item_list = GeospatialComputedItemList.load(
-            cast("List[Any]", resource.get("items")), cognite_client=cognite_client
+            cast(List[Any], resource.get("items")), cognite_client=cognite_client
         )
         return cls(item_list, cognite_client=cognite_client)