+++ cognite/client/data_classes/iam.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import json
 from typing import TYPE_CHECKING, Any, cast
 
 from typing_extensions import Self
@@ -48,8 +47,7 @@ class Group(CogniteResource):
         self._cognite_client = cast("CogniteClient", cognite_client)
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> Self:
-        resource = json.loads(resource) if isinstance(resource, str) else resource
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> Self:
         return cls(
             name=resource["name"],
             source_id=resource.get("sourceId"),
@@ -233,6 +231,5 @@ class ClientCredentials(CogniteResource):
         }
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> ClientCredentials:
-        resource = json.loads(resource) if isinstance(resource, str) else resource
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> ClientCredentials:
         return cls(client_id=resource["clientId"], client_secret=resource["clientSecret"])