+++ cognite/client/data_classes/templates.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import json
 from collections import UserDict
 from typing import TYPE_CHECKING, Any, ClassVar, Union, cast
 
@@ -264,31 +263,27 @@ class TemplateInstance(CogniteResource):
         }
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> TemplateInstance:
-        if isinstance(resource, str):
-            return cls.load(json.loads(resource), cognite_client=cognite_client)
-        elif isinstance(resource, dict):
-            instance = cls(cognite_client=cognite_client)
-            for key, value in resource.items():
-                snake_case_key = to_snake_case(key)
-                if hasattr(instance, snake_case_key):
-                    if key == "fieldResolvers":
-                        setattr(
-                            instance,
-                            snake_case_key,
-                            {
-                                key: TemplateInstance._field_resolver_load(field_resolver)
-                                for key, field_resolver in value.items()
-                            },
-                        )
-                    else:
-                        setattr(instance, snake_case_key, value)
-            return instance
-        raise TypeError(f"Resource must be json str or dict, not {type(resource)}")
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> TemplateInstance:
+        instance = cls(cognite_client=cognite_client)
+        for key, value in resource.items():
+            snake_case_key = to_snake_case(key)
+            if hasattr(instance, snake_case_key):
+                if key == "fieldResolvers":
+                    setattr(
+                        instance,
+                        snake_case_key,
+                        {
+                            key: TemplateInstance._field_resolver_load(field_resolver)
+                            for key, field_resolver in value.items()
+                        },
+                    )
+                else:
+                    setattr(instance, snake_case_key, value)
+        return instance
 
     @staticmethod
     def _field_resolver_load(resource: dict, cognite_client: CogniteClient | None = None) -> CogniteResource:
-        return TemplateInstance.field_resolver_mapper[resource["type"]].load(resource, cognite_client)
+        return TemplateInstance.field_resolver_mapper[resource["type"]]._load(resource, cognite_client)
 
 
 class TemplateInstanceUpdate(CogniteUpdate):
@@ -398,18 +393,14 @@ class View(CogniteResource):
             return value
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> View:
-        if isinstance(resource, str):
-            return cls.load(json.loads(resource), cognite_client=cognite_client)
-        elif isinstance(resource, dict):
-            instance = cls(cognite_client=cognite_client)
-            for key, value in resource.items():
-                snake_case_key = to_snake_case(key)
-                if hasattr(instance, snake_case_key):
-                    value = value if key != "source" else Source.load(value, cognite_client)
-                    setattr(instance, snake_case_key, value)
-            return instance
-        raise TypeError(f"Resource must be json str or dict, not {type(resource)}")
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> View:
+        instance = cls(cognite_client=cognite_client)
+        for key, value in resource.items():
+            snake_case_key = to_snake_case(key)
+            if hasattr(instance, snake_case_key):
+                value = value if key != "source" else Source._load(value, cognite_client)
+                setattr(instance, snake_case_key, value)
+        return instance
 
 
 class ViewResolveItem(UserDict, CogniteResource):
@@ -421,12 +412,8 @@ class ViewResolveItem(UserDict, CogniteResource):
         return self.data
 
     @classmethod
-    def load(cls, data: dict | str, cognite_client: CogniteClient | None = None) -> ViewResolveItem:
-        if isinstance(data, str):
-            return cls.load(json.loads(data), cognite_client=cognite_client)
-        elif isinstance(data, dict):
-            return cls(data, cognite_client=cognite_client)
-        raise TypeError(f"Resource must be json str or dict, not {type(data)}")
+    def _load(cls, data: dict[str, Any], cognite_client: CogniteClient | None = None) -> ViewResolveItem:
+        return cls(data, cognite_client=cognite_client)
 
 
 class GraphQlError(CogniteResource):
@@ -461,11 +448,10 @@ class GraphQlResponse(CogniteResource):
         return output
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> GraphQlResponse:
-        resource = json.loads(resource) if isinstance(resource, str) else resource
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> GraphQlResponse:
         return cls(
             data=resource.get("data"),
-            errors=[GraphQlError.load(error) for error in resource.get("errors", [])],
+            errors=[GraphQlError._load(error) for error in resource.get("errors", [])],
             cognite_client=cognite_client,
         )