+++ cognite/client/data_classes/transformations/jobs.py
@@ -43,11 +43,6 @@ class TransformationJobMetric(CogniteResource):
         self.count = count
         self._cognite_client = cast("CogniteClient", cognite_client)
 
-    @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> TransformationJobMetric:
-        instance = super().load(resource, cognite_client)
-        return instance
-
 
 class TransformationJobMetricList(CogniteResourceList[TransformationJobMetric]):
     _RESOURCE = TransformationJobMetric
@@ -259,8 +254,8 @@ class TransformationJob(CogniteResource):
         return output
 
     @classmethod
-    def load(cls, resource: dict | str, cognite_client: CogniteClient | None = None) -> TransformationJob:
-        instance = super().load(resource, cognite_client)
+    def _load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> TransformationJob:
+        instance = super()._load(resource, cognite_client)
         if isinstance(instance.destination, dict):
             instance.destination = _load_destination_dct(instance.destination)
         elif isinstance(instance.destination, str):