+++ CHANGELOG.md
@@ -48,7 +48,7 @@ with no easy way to add a prefix. Also, it no longer expands metadata by default
 - `client.data_modeling.instances.aggregate` returns `AggregatedNumberedValue | list[AggregatedNumberedValue] | InstanceAggregationResultList` depending
   on the `aggregates` and `group_by` parameters. Previously, it always returned `InstanceAggregationResultList`.
 - The `Group` attribute `capabilities` is now a `Capabilities` object, instead of a `dict`.
-
+- Support for `YAML` in all `CogniteResource.load()` and `CogniteResourceList.load()` methods. 
 
 ### Added
 - Added `load` implementation for `VisionResource`s: `ObjectDetection`, `TextRegion`, `AssetLink`, `BoundingBox`,